requirements.txt
streamlit
pandas
xlrd
pdfplumber
python-dotenv
pymysql
sqlalchemy
google-auth-oauthlib
google-api-python-client
PyJWT
streamlit-aggrid
extra-streamlit-components
xlsxwriter
openpyxl
email-to
secure-smtplib
streamlit-javascript
matplotlib
plotly
plotly-express
joblib
boto3
pycryptodome
razorpay
mysql-connector-python
Flask
streamlit_cookies_manager
dateutils==0.6.12
numpy==1.26.4
pandas==2.2.0
streamlit==1.37.0
XlsxWriter==3.2.0
proto-plus==1.23.0
protobuf==4.25.3
Pyrebase4==4.8.0
plotly
mysql-connector-python
google-auth-oauthlib
google-api-python-client
PyJWT
python-dotenv
plotly-express
streamlit-aggrid
Python-IO
extra-streamlit-components